cbor_encoder half typed arrays support

+++ include/jsoncons/json_reader.hpp
@@ -120,7 +120,7 @@ private:
         return other_handler_.uint64_value(value, tag, context, ec);
     }
 
-    bool do_half_value(double value, 
+    bool do_half_value(uint64_t value, 
                        semantic_tag tag,
                        const ser_context& context,
                        std::error_code& ec) override
